# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "绦管后台 API 文档",
    "description": "电商后台管理系统RESTful API，支持小程序对接",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://qimengzhiyue.cn/api/v1",
      "description": "生产环境"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "开发环境"
    }
  ],
...
  "paths": {
...
    "/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "分类列表",
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "刕efa分类",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "刕efa成功"
          }
        }
      }
    },
    "/categories/tree": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "分类树形结构",
        "parameters": [
          {
            "name": "flat",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "分类详情",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "更新分类",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "删除分类",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer Token 认证"
      }
    },
...
  }
}